+++ core/weights_reorder.h
@@ -22,15 +22,17 @@ namespace oidn {
 
   // Reorders weights from oihw to padded oihw format
   template<int K>
-  class WeightsReorder : public Node
+  class WeightsReorderNode : public Node
   {
   private:
     std::shared_ptr<memory> src;
     std::shared_ptr<memory> dst;
 
   public:
-    WeightsReorder(const std::shared_ptr<memory>& src, const std::shared_ptr<memory>& dst)
-      : src(src), dst(dst)
+    WeightsReorderNode(const std::shared_ptr<memory>& src,
+                       const std::shared_ptr<memory>& dst)
+      : src(src),
+        dst(dst)
     {
       memory::primitive_desc srcPrimDesc = src->get_primitive_desc();
       memory::primitive_desc dstPrimDesc = dst->get_primitive_desc();